# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AcknowledgeResponse": {
        "properties": {
          "count": {
            "default": 1,
            "title": "Count",
            "type": "integer"
          },
          "status": {
            "const": "acknowledged",
            "default": "acknowledged",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "AcknowledgeResponse",
        "type": "object"
      },
...
      "AlertHistoryPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertHistoryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before"
          }
        },
        "required": [
          "items",
          "next_before"
        ],
        "title": "AlertHistoryPage",
        "type": "object"
      },
      "AlertHistoryResponse": {
        "properties": {
          "acknowledged": {
            "title": "Acknowledged",
            "type": "boolean"
          },
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "integer"
          },
          "rule_name": {
            "title": "Rule Name",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          },
          "triggered_at": {
            "format": "date-time",
            "title": "Triggered At",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "id",
          "rule_id",
          "rule_name",
          "symbol",
          "priority",
          "message",
          "field",
          "value",
          "threshold",
          "acknowledged",
          "acknowledged_at",
          "triggered_at"
        ],
        "title": "AlertHistoryResponse",
        "type": "object"
      },
...
      "AlertRuleCreate": {
        "properties": {
          "alert_type": {
            "default": "threshold",
            "title": "Alert Type",
            "type": "string"
          },
          "comparison": {
            "enum": [
              "gt",
              "gte",
              "lt",
              "lte",
              "eq",
              "in"
            ],
            "title": "Comparison",
            "type": "string"
          },
          "condition_field": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Condition Field",
            "type": "string"
          },
          "cooldown_seconds": {
            "default": 30,
            "maximum": 86400.0,
            "minimum": 0.0,
            "title": "Cooldown Seconds",
            "type": "integer"
          },
          "message_template": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Template"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notify_discord": {
            "default": false,
            "title": "Notify Discord",
            "type": "boolean"
          },
          "notify_email": {
            "default": false,
            "title": "Notify Email",
            "type": "boolean"
          },
          "priority": {
            "default": "medium",
            "enum": [
              "critical",
              "high",
              "medium",
              "low",
              "info"
            ],
            "title": "Priority",
            "type": "string"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "name",
          "condition_field",
          "comparison",
          "threshold"
        ],
        "title": "AlertRuleCreate",
        "type": "object"
      },
      "AlertRuleResponse": {
        "properties": {
          "alert_type": {
            "title": "Alert Type",
            "type": "string"
          },
          "comparison": {
            "title": "Comparison",
            "type": "string"
          },
          "condition_field": {
            "title": "Condition Field",
            "type": "string"
          },
          "cooldown_seconds": {
            "title": "Cooldown Seconds",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_enabled": {
            "title": "Is Enabled",
            "type": "boolean"
          },
          "last_fired_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Fired At"
          },
          "message_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Template"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notify_discord": {
            "title": "Notify Discord",
            "type": "boolean"
          },
          "notify_email": {
            "title": "Notify Email",
            "type": "boolean"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "symbol",
          "alert_type",
          "condition_field",
          "comparison",
          "threshold",
          "priority",
          "message_template",
          "cooldown_seconds",
          "is_enabled",
          "notify_discord",
          "notify_email",
          "last_fired_at",
          "created_at",
          "updated_at"
        ],
        "title": "AlertRuleResponse",
        "type": "object"
      },
      "AlertRuleUpdate": {
        "properties": {
          "comparison": {
            "anyOf": [
              {
                "enum": [
                  "gt",
                  "gte",
                  "lt",
                  "lte",
                  "eq",
                  "in"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparison"
          },
          "condition_field": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition Field"
          },
          "cooldown_seconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldown Seconds"
          },
          "is_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Enabled"
          },
          "message_template": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Template"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notify_discord": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notify Discord"
          },
          "notify_email": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notify Email"
          },
          "priority": {
            "anyOf": [
              {
                "enum": [
                  "critical",
                  "high",
                  "medium",
                  "low",
                  "info"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          }
        },
        "title": "AlertRuleUpdate",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Real-time market-microstructure intelligence: streaming features, order-book analytics, calibrated ML signals, alerts and backtesting.",
    "title": "AlgoViz",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/alerts/history": {
      "get": {
        "operationId": "get_history_api_v1_alerts_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "`next_before` of the previous page",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`next_before` of the previous page",
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "unacknowledged_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unacknowledged Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertHistoryPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Alert history, newest first",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/v1/alerts/history/acknowledge-all": {
      "post": {
        "operationId": "acknowledge_all_api_v1_alerts_history_acknowledge_all_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcknowledgeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Acknowledge all alerts",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/v1/alerts/history/{alert_id}/acknowledge": {
      "post": {
        "operationId": "acknowledge_alert_api_v1_alerts_history__alert_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcknowledgeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Acknowledge alert",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/v1/alerts/rules": {
      "get": {
        "operationId": "list_rules_api_v1_alerts_rules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertRuleResponse"
                  },
                  "title": "Response List Rules Api V1 Alerts Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "List alert rules",
        "tags": [
          "Alerts"
        ]
      },
      "post": {
        "operationId": "create_rule_api_v1_alerts_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create alert rule",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/v1/alerts/rules/{rule_id}": {
      "delete": {
        "operationId": "delete_rule_api_v1_alerts_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete alert rule",
        "tags": [
          "Alerts"
        ]
      },
      "patch": {
        "operationId": "update_rule_api_v1_alerts_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRuleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update alert rule",
        "tags": [
          "Alerts"
        ]
      }
    },
...
  }
}